extract table hover test to desktop specific file
Mobile does not support hovering and therefore looking
into `table.hoveredCell.Col` would crash.
