+++ widget/table_test.go
@@ -104,44 +104,6 @@ func TestTable_Unselect(t *testing.T) {
 	test.AssertImageMatches(t, "table/theme_initial.png", w.Canvas().Capture())
 }
 
-func TestTable_Hovered(t *testing.T) {
-	app := test.NewApp()
-	defer test.NewApp()
-	app.Settings().SetTheme(theme.LightTheme())
-
-	table := NewTable(
-		func() (int, int) { return 2, 2 },
-		func() fyne.CanvasObject {
-			return NewLabel("placeholder")
-		},
-		func(id TableCellID, c fyne.CanvasObject) {
-			c.(*Label).SetText(fmt.Sprintf("Cell %d, %d", id.Row, id.Col))
-		})
-
-	w := test.NewWindow(table)
-	defer w.Close()
-	w.Resize(fyne.NewSize(180, 180))
-
-	test.MoveMouse(w.Canvas(), fyne.NewPos(35, 50))
-	test.MoveMouse(w.Canvas(), fyne.NewPos(35, 100))
-
-	assert.Nil(t, table.hoveredCell)
-
-	test.AssertImageMatches(t, "table/hovered_out.png", w.Canvas().Capture())
-
-	table.Length = func() (int, int) { return 3, 5 }
-	table.Refresh()
-
-	w.SetContent(table)
-	w.Resize(fyne.NewSize(180, 180))
-	test.MoveMouse(w.Canvas(), fyne.NewPos(35, 50))
-
-	assert.Equal(t, 0, table.hoveredCell.Col)
-	assert.Equal(t, 1, table.hoveredCell.Row)
-
-	test.AssertImageMatches(t, "table/hovered.png", w.Canvas().Capture())
-}
-
 func TestTable_Refresh(t *testing.T) {
 	displayText := "placeholder"
 	table := NewTable(
